Changes in credit note and debit note

+++ DataLayer/Service/ACTransactionsService.cs
@@ -740,7 +740,7 @@ namespace MicroApi.DataLayer.Service
                         dt.Columns.Add("AMOUNT", typeof(double));
                         dt.Columns.Add("VAT_AMOUNT", typeof(double));
                         dt.Columns.Add("REMARKS", typeof(string));
-                        dt.Columns.Add("GST_PERC", typeof(float));
+                        dt.Columns.Add("GST_PERC", typeof(int));
                         dt.Columns.Add("HSN_CODE", typeof(string));
                         dt.Columns.Add("CGST", typeof(decimal));
                         dt.Columns.Add("SGST", typeof(decimal));
@@ -832,7 +832,7 @@ namespace MicroApi.DataLayer.Service
                         dt.Columns.Add("AMOUNT", typeof(double));
                         dt.Columns.Add("VAT_AMOUNT", typeof(double));
                         dt.Columns.Add("REMARKS", typeof(string));
-                        dt.Columns.Add("GST_PERC", typeof(float));
+                        dt.Columns.Add("GST_PERC", typeof(int));
                         dt.Columns.Add("HSN_CODE", typeof(string));
                         dt.Columns.Add("CGST", typeof(decimal));
                         dt.Columns.Add("SGST", typeof(decimal));
@@ -1009,7 +1009,7 @@ namespace MicroApi.DataLayer.Service
                                     AMOUNT = reader["AMOUNT"] != DBNull.Value ? Convert.ToSingle(reader["AMOUNT"]) : 0,
                                     GST_AMOUNT = reader["VAT_AMOUNT"] != DBNull.Value ? Convert.ToSingle(reader["VAT_AMOUNT"]) : 0,
                                     REMARKS = reader["REMARKS"]?.ToString(),
-                                    GST_PERC = reader["GST_PERC"] != DBNull.Value ? Convert.ToSingle(reader["GST_PERC"]) : 0,
+                                    GST_PERC = reader["GST_PERC"] != DBNull.Value ? Convert.ToInt32(reader["GST_PERC"]) : 0,
                                     HSN_CODE = reader["HSN_CODE"]?.ToString(),
                                     CGST = reader["CGST"] != DBNull.Value ? Convert.ToDecimal(reader["CGST"]) : 0,
                                     SGST = reader["SGST"] != DBNull.Value ? Convert.ToDecimal(reader["SGST"]) : 0
@@ -1103,7 +1103,7 @@ namespace MicroApi.DataLayer.Service
                         dt.Columns.Add("AMOUNT", typeof(double));
                         dt.Columns.Add("VAT_AMOUNT", typeof(double));
                         dt.Columns.Add("REMARKS", typeof(string));
-                        dt.Columns.Add("GST_PERC", typeof(float));
+                        dt.Columns.Add("GST_PERC", typeof(int));
                         dt.Columns.Add("HSN_CODE", typeof(string));
                         dt.Columns.Add("CGST", typeof(decimal));
                         dt.Columns.Add("SGST", typeof(decimal));
