+++ Models/AC_CreditNote.cs
@@ -49,7 +49,7 @@
         public double AMOUNT { get; set; }
         public double GST_AMOUNT { get; set; }
         public string REMARKS { get; set; }
-        public float? GST_PERC { get; set; }
+        public int? GST_PERC { get; set; }
         public string? HSN_CODE { get; set; }
         public decimal? CGST { get; set; }
         public decimal? SGST { get; set; }
@@ -112,7 +112,7 @@
         public string REMARKS { get; set; }
         public decimal? CGST { get; set; }
         public decimal? SGST { get; set; }
-        public float? GST_PERC { get; set; }//FOR SELECT
+        public int? GST_PERC { get; set; }//FOR SELECT
         public string? HSN_CODE { get; set; }
         public string? LEDGER_CODE { get; set; }
         public string? LEDGER_NAME { get; set; }
